refactor(ramps): move Money Account wallet signing from kyc to ramps
Wallet ownership signing is a Money Movement / neobank-proxy concern, not
KYC. Moving it into ramps-controller also removes the ramps→kyc package
dependency that broke the monorepo project-reference build and blocked
publish-preview for core PR #9853.
