+++ packages/kyc-controller/package.json
@@ -61,7 +61,6 @@
     "@metamask/base-data-service": "^0.1.3",
     "@metamask/controller-utils": "^12.3.0",
     "@metamask/geolocation-controller": "^1.0.0",
-    "@metamask/keyring-controller": "^27.1.1",
     "@metamask/messenger": "^2.0.0",
     "@metamask/profile-sync-controller": "^29.0.0",
     "@metamask/superstruct": "^3.4.1",
