+++ packages/kyc-controller/src/KycController-method-action-types.ts
@@ -216,26 +216,6 @@ export type KycControllerGetSessionStatusAction = {
 };
 
 /**
- * Registers a Money Account wallet with MoonPay Iron via neobank-proxy.
- *
- * Consumers provide only the Monad address. The controller reuses the Iron
- * customer id captured from MoonPay's hosted frame when available, otherwise
- * it resolves the id via `GET /neobank/customers/{external_id}/external`
- * (MetaMask canonical profile id). Customer resolution happens before the
- * first list/lookup because list requires `customer_id` in the path.
- * Message construction, signing, submission, and ambiguous-write
- * reconciliation stay internal to KYC.
- *
- * @param params - Money Account wallet registration parameters.
- * @param params.address - Monad Money Account address.
- * @returns The successful registration state.
- */
-export type KycControllerRegisterMoneyAccountWalletAction = {
-  type: `KycController:registerMoneyAccountWallet`;
-  handler: KycController['registerMoneyAccountWallet'];
-};
-
-/**
  * Resets the flow to idle, clearing session tokens and sub-flow state while
  * preserving persisted terms acceptance and the per-product cache.
  */
@@ -263,5 +243,4 @@ export type KycControllerMethodActions =
   | KycControllerStartSumSubAction
   | KycControllerRefreshKycStatusAction
   | KycControllerGetSessionStatusAction
-  | KycControllerRegisterMoneyAccountWalletAction
   | KycControllerResetAction;
