+++ packages/kyc-controller/src/index.ts
@@ -8,7 +8,6 @@ export type {
   KycControllerEvents,
   KycControllerGetStateAction,
   KycControllerMessenger,
-  MoneyAccountWalletRegistrationResult,
   KycControllerOptions,
   KycControllerState,
   KycControllerStateChangeEvent,
@@ -30,7 +29,6 @@ export type {
   KycControllerLoadDisclaimersAction,
   KycControllerRefreshKycStatusAction,
   KycControllerResetAction,
-  KycControllerRegisterMoneyAccountWalletAction,
   KycControllerStartSumSubAction,
 } from './KycController-method-action-types.js';
 
@@ -42,7 +40,6 @@ export type {
   CreateSessionParams,
   CreateUkycSessionParams,
   GetSessionStatusParams,
-  GetWalletRegistrationStatusParams,
   GetWrappingKeyParams,
   IronCustomerResponse,
   JwksResponse,
@@ -53,7 +50,6 @@ export type {
   KycServiceInvalidateQueriesAction,
   KycServiceMessenger,
   KycServiceOptions,
-  RegisterSelfHostedWalletParams,
   SubmitConsentsParams,
   UkycSessionResponse,
   WrappedEncryptionKey,
@@ -71,11 +67,8 @@ export type {
   KycServiceFetchJwksAction,
   KycServiceFetchKycStatusAction,
   KycServiceGetGeoCountryAction,
-  KycServiceGetMoonpayCustomerIdAction,
   KycServiceGetSessionStatusAction,
-  KycServiceGetWalletRegistrationStatusAction,
   KycServiceGetWrappingKeyAction,
-  KycServiceRegisterSelfHostedWalletAction,
   KycServiceSubmitConsentsAction,
 } from './KycService-method-action-types.js';
 
@@ -141,9 +134,3 @@ export type {
   MintedUkycTestToken,
   MintUkycTestTokenParams,
 } from './ukyc/testToken.js';
-
-export type {
-  SelfHostedRegistration,
-  WalletRegistrationErrorKind,
-} from './wallet-registration-service.js';
-export { WalletRegistrationError } from './wallet-registration-service.js';
