+++ packages/kyc-controller/tsconfig.build.json
@@ -10,7 +10,6 @@
     { "path": "../base-data-service/tsconfig.build.json" },
     { "path": "../controller-utils/tsconfig.build.json" },
     { "path": "../geolocation-controller/tsconfig.build.json" },
-    { "path": "../keyring-controller/tsconfig.build.json" },
     { "path": "../messenger/tsconfig.build.json" },
     { "path": "../profile-sync-controller/tsconfig.build.json" }
   ],
