+++ packages/kyc-controller/tsconfig.json
@@ -8,7 +8,6 @@
     { "path": "../base-data-service" },
     { "path": "../controller-utils" },
     { "path": "../geolocation-controller" },
-    { "path": "../keyring-controller" },
     { "path": "../messenger" },
     { "path": "../profile-sync-controller" }
   ],
