+++ packages/ramps-controller/src/wallet-registration-machine.test.ts
