+++ packages/ramps-controller/src/wallet-registration-machine.ts
