+++ packages/ramps-controller/src/wallet-registration-service.test.ts
