+++ packages/ramps-controller/src/wallet-registration-service.ts
