+++ packages/ramps-controller/tsconfig.build.json
@@ -21,9 +21,6 @@
     },
     {
       "path": "../remote-feature-flag-controller/tsconfig.build.json"
-    },
-    {
-      "path": "../kyc-controller/tsconfig.build.json"
     }
   ],
   "include": ["../../types", "./src"]
