+++ packages/ramps-controller/tsconfig.json
@@ -19,9 +19,6 @@
     },
     {
       "path": "../controller-utils"
-    },
-    {
-      "path": "../kyc-controller"
     }
   ],
   "include": ["../../types", "./src"]
